{colombo} merging with trunk
git-svn-id: http://svn.code.sf.net/p/sumo/code/branches/colombo@14457 29f149b2-0c63-11e1-8f5c-a347f6a9acd0

+++ src/utils/geom/PositionVector.cpp
@@ -9,7 +9,7 @@
 ///
 // A list of positions
 /****************************************************************************/
-// SUMO, Simulation of Urban MObility; see http://sumo.sourceforge.net/
+// SUMO, Simulation of Urban MObility; see http://sumo-sim.org/
 // Copyright (C) 2001-2013 DLR (http://www.dlr.de/) and contributors
 /****************************************************************************/
 //
@@ -1040,7 +1040,7 @@ PositionVector::insertAt(int index, const Position& p) {
 void
 PositionVector::replaceAt(int index, const Position& p) {
     assert(index < static_cast<int>(size()));
-    assert(index + size() >= 0);
+    assert(index + static_cast<int>(size()) >= 0);
     if (index >= 0) {
         (*this)[index] = p;
     } else {
